+++ CHANGELOG.md
@@ -28,6 +28,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Implement `Index` and `IndexMut` for `ua::Array` to allow direct element access.
 - Add methods `Server::write_object_property()` and `Server::read_object_property()`.
 - Add methods `Server::create_event()` and `Server::trigger_event()`.
+- Add `Lifecycle` and `LifecyclManager`.
 
 ### Changed